docs: Go house style guide + lint hardening (#927)
Adds docs/go-style.md — the repo's Go house style: rationale-only
comments citing spec/tool/device, operator-sentence errors, the
honesty doctrine (no silent fallbacks; tests assert from the
authority), and the linting policy (a linter is only enabled when
the whole tree already passes it).

Enables misspell, unconvert, wastedassign in .golangci.yml and
makes the tree pass them: 19 redundant conversions removed and 4
wasted assignments dropped across cmd/dhs, acp2, amwa, cerebrum-nb,
emberplus, osc, probel-sw02p, export.

+++ internal/acp2/consumer/plugin.go
@@ -471,7 +471,7 @@ func (p *Plugin) GetValue(ctx context.Context, req consumer.ValueRequest) (consu
 
 	// req.PID overrides pid=8 (value); req.Idx overrides idx=0 (active).
 	// Zero defaults preserve historical behaviour.
-	targetPID := uint8(codec.PIDValue)
+	targetPID := codec.PIDValue
 	if req.PID > 0 {
 		targetPID = uint8(req.PID)
 	}
